requirement.txt
decord==0.6.0
diffusers==0.35.0
einops==0.8.0
huggingface-hub==0.34.4
hydra-core==1.3.2
matplotlib==3.10.0
numpy==1.26.3
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.2
Pillow==9.5.0
requests==2.32.3
safetensors==0.4.5
sam2==1.1.0
scipy==1.14.1
tqdm==4.67.1
transformers==4.57.0
sentencepiece
accelerate
timm
